+++ src/test/java/ro/ubb/repository/db/DBStudentRepositoryTest.java
@@ -12,7 +12,7 @@ import java.util.HashSet;
 import java.util.Optional;
 import java.util.Set;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class DBStudentRepositoryTest {
   private Repository<Long, Student> studentRepository;
@@ -21,33 +21,32 @@ class DBStudentRepositoryTest {
   @BeforeEach
   void setUp() {
     studentRepository =
-            new DBStudentRepository(
-                    "configuration" + FileSystems.getDefault().getSeparator() + "db-credentials.data",
-                    "public.\"Students_test\"");
+        new DBStudentRepository(
+            "configuration" + FileSystems.getDefault().getSeparator() + "db-credentials.data",
+            "public.\"Students_test\"");
     student = new Student("sn1", "studentName", 1);
     student.setId(1L);
     studentRepository.save(student);
-
   }
 
   @AfterEach
   void tearDown() {
-    if(studentRepository.findOne(1L).isPresent())
-      studentRepository.delete(1L);
-    student =null;
+    if (studentRepository.findOne(1L).isPresent()) studentRepository.delete(1L);
+    student = null;
     studentRepository = null;
-
   }
+
   @Test
   void
-  Given_StudentRepositoryWithOneEntity_When_FindingSerialNumberOfOneThatIsInRepository_Then_ReturnSoughtSerialNumber() {
+      Given_StudentRepositoryWithOneEntity_When_FindingSerialNumberOfOneThatIsInRepository_Then_ReturnSoughtSerialNumber() {
     Assertions.assertEquals("sn1", studentRepository.findOne(1L).get().getSerialNumber());
   }
 
   @Test
   void
-  Given_StudentRepositoryWithOneEntity_When_FindingOneThatIsNotInRepository_Then_ReturnFalse() {
-    Assertions.assertFalse(studentRepository.findOne(2L).isPresent(), "member with id given not found");
+      Given_StudentRepositoryWithOneEntity_When_FindingOneThatIsNotInRepository_Then_ReturnFalse() {
+    Assertions.assertFalse(
+        studentRepository.findOne(2L).isPresent(), "member with id given not found");
   }
 
   @Test
@@ -57,13 +56,13 @@ class DBStudentRepositoryTest {
 
   @Test
   void
-  Given_StudentRepositoryWithOneEntity_WhenFindOneWithNullId_Then_ThrowIllegalArgumentException() {
+      Given_StudentRepositoryWithOneEntity_WhenFindOneWithNullId_Then_ThrowIllegalArgumentException() {
     Assertions.assertThrows(IllegalArgumentException.class, () -> studentRepository.findOne(null));
   }
 
   @Test
   void
-  Given_StudentRepositoryWithOneEntity_When_FindingAllInRepository_Then_ReturnSetWithThatEntity() {
+      Given_StudentRepositoryWithOneEntity_When_FindingAllInRepository_Then_ReturnSetWithThatEntity() {
     Set<Student> students = new HashSet<>();
     students.add(student);
     Assertions.assertEquals(students, studentRepository.findAll());
@@ -71,7 +70,7 @@ class DBStudentRepositoryTest {
 
   @Test
   void
-  Given_StudentRepositoryWithOneEntity_When_SavingNewValidEntity_Then_NumberOfEntitiesInRepositoryIncrements() {
+      Given_StudentRepositoryWithOneEntity_When_SavingNewValidEntity_Then_NumberOfEntitiesInRepositoryIncrements() {
     Student student2 = new Student("sn2", "studentName2", 2);
     student2.setId(2L);
     studentRepository.save(student2);
@@ -81,7 +80,7 @@ class DBStudentRepositoryTest {
 
   @Test
   void
-  Given_StudentRepositoryWithOneEntity_When_SavingEntityWithIdAlreadyInRepository_Then_SizeOfRepositoryRemainsOne() {
+      Given_StudentRepositoryWithOneEntity_When_SavingEntityWithIdAlreadyInRepository_Then_SizeOfRepositoryRemainsOne() {
     studentRepository.save(student);
     Assertions.assertEquals(1, ((Set<Student>) studentRepository.findAll()).size());
   }
@@ -93,20 +92,21 @@ class DBStudentRepositoryTest {
 
   @Test
   void
-  Given_StudentRepositoryWithOneEntity_When_DeletingSingleEntityInRepository_Then_SizeOfRepositoryIsZero() {
+      Given_StudentRepositoryWithOneEntity_When_DeletingSingleEntityInRepository_Then_SizeOfRepositoryIsZero() {
     studentRepository.delete(1L);
     Assertions.assertEquals(0, ((Set<Student>) studentRepository.findAll()).size());
   }
 
   @Test
   void
-  Given_StudentRepositoryWithOneEntity_When_DeletingEntityInRepository_Then_ReturnRightDataAboutTheDeletedEntity() {
+      Given_StudentRepositoryWithOneEntity_When_DeletingEntityInRepository_Then_ReturnRightDataAboutTheDeletedEntity() {
     Assertions.assertEquals("sn1", studentRepository.delete(1L).get().getSerialNumber());
   }
 
   @Test
   void Given_StudentRepositoryWithOneEntity_When_DeletingEntityNotInRepository_Then_ReturnFalse() {
-    Assertions.assertFalse(studentRepository.delete(2L).isPresent(), "member with id given not found");
+    Assertions.assertFalse(
+        studentRepository.delete(2L).isPresent(), "member with id given not found");
   }
 
   @Test
@@ -116,13 +116,13 @@ class DBStudentRepositoryTest {
 
   @Test
   void
-  Given_StudentRepositoryWithOneEntity_When_DeletingWithNullId_Then_ThrowsIllegalArgumentException() {
+      Given_StudentRepositoryWithOneEntity_When_DeletingWithNullId_Then_ThrowsIllegalArgumentException() {
     Assertions.assertThrows(IllegalArgumentException.class, () -> studentRepository.delete(null));
   }
 
   @Test
   void
-  Given_StudentRepositoryWithOneEntity_When_UpdatingEntityNotInRepository_Then_ReturnOptionalWithEntity() {
+      Given_StudentRepositoryWithOneEntity_When_UpdatingEntityNotInRepository_Then_ReturnOptionalWithEntity() {
     Student updatedStudent2 = new Student("sn2", "updatedStudentName2", 2);
     updatedStudent2.setId(2L);
     Optional<Student> updateResult = studentRepository.update(updatedStudent2);
@@ -140,7 +140,7 @@ class DBStudentRepositoryTest {
 
   @Test
   void
-  Given_StudentRepositoryWithOneEntity_When_UpdatingEntityInRepository_Then_ReturnOptionalWithNoValue() {
+      Given_StudentRepositoryWithOneEntity_When_UpdatingEntityInRepository_Then_ReturnOptionalWithNoValue() {
     Student updatedStudent = new Student("sn1", "updatedStudentName", 1);
     updatedStudent.setId(1L);
     Assertions.assertFalse(studentRepository.update(updatedStudent).isPresent());
@@ -148,7 +148,7 @@ class DBStudentRepositoryTest {
 
   @Test
   void
-  Given_StudentRepositoryWithOneEntity_When_UpdatingWithNullId_Then_ThrowsIllegalArgumentException() {
+      Given_StudentRepositoryWithOneEntity_When_UpdatingWithNullId_Then_ThrowsIllegalArgumentException() {
     Assertions.assertThrows(IllegalArgumentException.class, () -> studentRepository.update(null));
   }
 }
